Improve e2e tests structure
