+++ docs/e2e-test.md
@@ -4,12 +4,32 @@ This doc gives short instructions how to run e2e tests. For the developing e2e t
 
 ## Prerequisites
 
-- Make sure that `make` and `go` are available
-- Make sure that 'docker run' works without sudo, you can check it with
+1. Make sure that `make` is available
 
-  ```sh
-  docker run hello-world
-  ```
+   To install make:
+   - Ubuntu
+
+   ```sh
+   sudo apt install build-essential
+   ```
+
+   - Centos
+
+   ```sh
+   sudo yum groupinstall "Development Tools"
+   ```
+
+1. Make sure that 'docker run' works without sudo, you can check it with
+
+   ```sh
+   docker ps
+   ```
+
+   Else add the user to `docker` group:
+
+   ```sh
+   sudo groupadd docker && sudo usermod -aG docker $USER && newgrp docker
+   ```
 
 ## Running
 
@@ -29,24 +49,48 @@ make clean -C /opt/metal3-dev-env/metal3-dev-env/ && sudo rm -rf /opt/metal3-dev
 
 ## Included tests
 
-The e2e tests currently run one after the other, using the same environment for all of them.
-They run in this order:
+The e2e tests currently include three different sets:
 
-1. Remediation
-1. Inspection¹
-1. Pivoting
-1. Upgrade BMO
-1. Upgrade Ironic
-1. Certificate rotation
-1. Node reuse
-1. Re-pivoting
+1. pivoting_based_feature_tests
+1. remediation_feature_tests
+1. upgrade_tests
+
+`pivoting_based_feature_tests`: Because these tests run mainly in the target cluster, they are dependent on the pivoting test and need to run in the following order:
+
+- Pivoting
+- Certificate rotation
+- Node reuse
+- Re-pivoting
+
+However, in case we need to run them in the ephemeral cluster pivoting and re-pivoting should be ignored.
+
+`remediation_feature_tests`: independent from the previous tests and can run independently includes:
+
+- Remediation
+- Inspection¹
 
 ¹ Inspection is actually run [in the middle of the remediation test](https://github.com/metal3-io/cluster-api-provider-metal3/blob/8d08f375de93a793f839b42b5ec40e6bebf98664/test/e2e/remediation_test.go#L108) for practical reasons at the moment.
 
 The ephemeral cluster is first launched using [metal3-dev-env](https://github.com/metal3-io/metal3-dev-env).
-The remediation and inspection tests are then run with the controllers still in the ephemeral cluster.
-After this, the controllers are moved to the target cluster during pivoting.
-From this point on all tests are run in the target cluster, until the controllers are moved back to the ephemeral cluster during re-pivoting.
+The remediation and inspection tests are then run with the controllers still in the ephemeral cluster either before pivoting or after re-pivoting.
+
+`upgrade_management_cluster_test`:
+
+- Upgrade BMO
+- Upgrade Ironic
+- Upgrade CAPI/CAPM3
+
+Guidelines to follow when adding new E2E tests:
+
+- Tests should be defined in a new file and separate test spec, unless the new test depends on existing tests.
+- Tests are categorized with a custom label that can be used to filter a set of E2E tests to be run. To run a subset of tests, a combination of either one or both of the `GINKGO_FOCUS` and `GINKGO_SKIP` env variables can be set. The labels are defined in the description of the `Describe` container between `[]`, for example:
+
+`[upgrade]` => runs only existing upgrade tests including CAPI, CAPM3, Ironic and Baremetal Operator.
+`[remediation]` => runs only remediation and inspection tests.
+
+For instance, to skip the upgrade E2E tests set `GINKGO_SKIP="[upgrade]"`
+
+`GINKGO_FOCUS` and `GINKGO_SKIP` are defined in the Makefile, and should be initialized in the CI JJBs to select the target e2e tests.
 
 ### Note on BMO and Ironic upgrade
